Backend/requirements.txt
fastapi
uvicorn
pydantic
openai
groq
python-dotenv
httpx
